move schedules to it's own file
