+++ tests/test_promises.py
@@ -15,7 +15,7 @@ import pytest
 from resonate.codec import Codec, NoopEncryptor
 from resonate.error import ServerError
 from resonate.network import LocalNetwork
-from resonate.promises import Promises, Schedules
+from resonate.promises import Promises
 from resonate.send import Sender
 from resonate.transport import Transport
 from resonate.types import Value
@@ -23,17 +23,17 @@ from resonate.types import Value
 I64_MAX = 2**63 - 1
 
 
-def _local() -> tuple[Promises, Schedules]:
+def _local() -> Promises:
     """Build promise/schedule clients sharing one local network, like ``Resonate::local()``."""
     net = LocalNetwork(pid="default", group="default")
     sender = Sender(Transport(net), None)
     codec = Codec(NoopEncryptor())
-    return Promises(sender, codec), Schedules(sender, codec)
+    return Promises(sender, codec)
 
 
 @pytest.mark.asyncio
 async def test_promises_create_get_resolve_roundtrip() -> None:
-    promises, _ = _local()
+    promises = _local()
 
     created = await promises.create("unit-p1", I64_MAX, Value(data={"x": 1}), {})
     assert created.id == "unit-p1"
@@ -53,7 +53,7 @@ async def test_promises_create_get_resolve_roundtrip() -> None:
 async def test_promises_create_get_reject_roundtrip() -> None:
     # Symmetric counterpart to the resolve roundtrip: ``reject`` settles the
     # promise as ``rejected`` (via ``_settle``), and the state survives a re-get.
-    promises, _ = _local()
+    promises = _local()
 
     created = await promises.create("unit-p-reject", I64_MAX, Value(data={"x": 1}), {})
     assert created.state == "pending"
@@ -67,45 +67,6 @@ async def test_promises_create_get_reject_roundtrip() -> None:
 
 @pytest.mark.asyncio
 async def test_promises_get_missing_returns_server_error() -> None:
-    promises, _ = _local()
+    promises = _local()
     with pytest.raises(ServerError):
         await promises.get("does-not-exist")
-
-
-@pytest.mark.asyncio
-async def test_schedules_create_get_delete_roundtrip() -> None:
-    _, schedules = _local()
-
-    created = await schedules.create(
-        "unit-s1", "*/5 * * * *", "unit-s1.{{.timestamp}}", 60_000, Value()
-    )
-    assert created.id == "unit-s1"
-    assert created.cron == "*/5 * * * *"
-
-    fetched = await schedules.get("unit-s1")
-    assert fetched.id == "unit-s1"
-
-    await schedules.delete("unit-s1")
-
-
-@pytest.mark.asyncio
-async def test_schedules_delete_missing_returns_server_error() -> None:
-    _, schedules = _local()
-    with pytest.raises(ServerError):
-        await schedules.delete("no-such-schedule")
-
-
-@pytest.mark.asyncio
-async def test_schedules_search_returns_record() -> None:
-    _, schedules = _local()
-
-    await schedules.create(
-        "unit-s-search",
-        "* * * * *",
-        "unit-s-search.{{.timestamp}}",
-        60_000,
-        Value(),
-    )
-
-    result = await schedules.search(None, 100, None)
-    assert any(s.id == "unit-s-search" for s in result.schedules)
